Minor tangent, port search_utils to ts (had started to rewrite SearchConfigTypeahead in TS, but I'll leave that off for later)
