--- client/src/search/search_utils.js
@@ -1,143 +1,0 @@
-import { gql, useQuery } from "@apollo/client";
-import _ from "lodash";
-import React from "react";
-
-import { lang } from "src/core/injected_build_constants";
-
-import { escapeRegExp } from "src/general_utils";
-
-const query_to_regexps_func = (query) => {
-  const raw_tokens = _.filter(query.split(" "));
-  const reg_exps = _.map(
-    raw_tokens,
-    (token) => new RegExp(`${escapeRegExp(_.deburr(token))}(?![^<]*>)`, "gi")
-  );
-  return reg_exps;
-};
-const query_to_reg_exps = _.memoize(query_to_regexps_func);
-
-// Used where Highlighter component can't be, e.g. where searched string already
-// contains markup and will need to be rendered with dangerouslySetInnerHTML
-const highlight_search_match = (search, content) => {
-  const reg_exps = query_to_reg_exps(search);
-
-  let modified_string = _.clone(content);
-  _.each(
-    reg_exps,
-    (reg_exp) =>
-      (modified_string = modified_string.replace(
-        reg_exp,
-        (match) => `<strong>${match}</strong>`
-      ))
-  );
-
-  return modified_string;
-};
-
-const split_matched_search_tokens = (search, content) => {
-  const reg_exps = query_to_reg_exps(search);
-
-  const split_token = "Ø";
-
-  let modified_string = _.clone(content);
-  _.each(
-    reg_exps,
-    (reg_exp) =>
-      (modified_string = modified_string.replace(
-        reg_exp,
-        (match) => `${split_token}${match}${split_token}`
-      ))
-  );
-
-  const split_string = _.split(modified_string, split_token);
-
-  return split_string;
-};
-const format_data = (
-  name_function,
-  menu_content_function,
-  data,
-  config_group_name
-) => ({
-  data,
-  name: name_function(data),
-  menu_content: (search) =>
-    _.isFunction(menu_content_function) ? (
-      menu_content_function(data, search, name_function)
-    ) : (
-      <InfoBaseHighlighter search={search} content={name_function(data)} />
-    ),
-  config_group_name,
-});
-
-const get_gql_query = (query_value, gql_search_configs) => gql`
-  query {
-    root(lang: "${lang}") {
-      ${_.reduce(
-        gql_search_configs,
-        (query_result, { query }) => `
-      ${query_result}
-      ${query(query_value)}
-      `,
-        `non_field`
-      )}
-    }
-  }
-`;
-const useSearchQuery = (query_value, gql_search_configs) => {
-  const query = get_gql_query(query_value, gql_search_configs);
-  const res = useQuery(query, {
-    skip: query_value.length === 0 || gql_search_configs.length === 0,
-  });
-  if (!res.loading && res.data) {
-    const data = _.flatMap(
-      gql_search_configs,
-      ({
-        queried_data_accessor,
-        name_function,
-        menu_content_function,
-        config_name,
-      }) => {
-        return _.map(res.data.root[queried_data_accessor], (row) =>
-          format_data(name_function, menu_content_function, row, config_name)
-        );
-      }
-    );
-
-    return { ...res, data };
-  }
-  return res;
-};
-
-class InfoBaseHighlighter extends React.Component {
-  render() {
-    const { search, content } = this.props;
-
-    const string_split_on_matched_tokens = split_matched_search_tokens(
-      search,
-      content
-    );
-
-    return (
-      <span>
-        {_.map(string_split_on_matched_tokens, (sub_string, ix) =>
-          ix % 2 === 0 ? (
-            <span key={ix}>{sub_string}</span>
-          ) : (
-            <mark key={ix} className="typeahead__highlight">
-              {sub_string}
-            </mark>
-          )
-        )}
-      </span>
-    );
-  }
-}
-
-export {
-  query_to_reg_exps,
-  highlight_search_match,
-  format_data,
-  useSearchQuery,
-  InfoBaseHighlighter,
-};
